requirements.txt
# Core
PySide6>=6.6.0
Pillow>=10.0.0

# Optional — RAW support
rawpy>=0.19.0

# Optional — FITS support
astropy>=6.0.0
# fitsio>=1.2.0  # faster C backend, uncomment if needed
numpy>=1.26.0

# Optional — PSD support
psd-tools>=1.9.0

# Optional — HEIC/HEIF support
pillow-heif>=0.16.0

# Optional — memory stats
psutil>=5.9.0

# Metadata editing — lossless JPEG EXIF write
piexif>=1.1.3
